new/requirement.txt
attrs==19.3.0
backports-abc==0.5
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
bleach==3.1.3
click==7.1.1
configparser==4.0.2
contextlib2==0.6.0.post1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
Django==1.11.29
django-rest-auth==0.9.5
djangorestframework==3.9.4
entrypoints==0.3
enum34==1.1.10
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
functools32==3.2.3.post2
futures==3.3.0
importlib-metadata==1.5.0
ipaddress==1.0.23
ipykernel==4.10.1
ipython==5.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==5.2.0
jupyter-core==4.6.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==2.2.5
mistune==0.8.4
mysqlclient==1.3.10
nbconvert==5.6.1
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.6
pandocfilters==1.4.2
pathlib2==2.3.5
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==1.0.18
psycopg2==2.7.4
psycopg2-binary==2.8.5
ptyprocess==0.6.0
Pygments==2.5.2
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
scandir==1.10.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.14.0
SQLAlchemy==1.3.15
subprocess32==3.5.4
terminado==0.8.3
testpath==0.4.4
tornado==5.1.1
traitlets==4.3.3
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
zipp==1.2.0
